+++ client/browser/src/libs/phabricator/extension.tsx
@@ -4,9 +4,12 @@ import { Observable } from 'rxjs'
 import { startWith } from 'rxjs/operators'
 import { setSourcegraphUrl } from '../../shared/util/context'
 import { MutationRecordLike, observeMutations } from '../../shared/util/dom'
-import { injectCodeIntelligence } from '../code_intelligence'
+import { determineCodeHost, injectCodeIntelligenceToCodeHost } from '../code_intelligence'
 import { getPhabricatorCSS, getSourcegraphURLFromConduit } from './backend'
 import { metaClickOverride } from './util'
+
+// Just for informational purposes (see getPlatformContext())
+window.SOURCEGRAPH_PHABRICATOR_EXTENSION = true
 
 // NOT idempotent.
 async function injectModules(): Promise<void> {
@@ -24,7 +27,10 @@ async function injectModules(): Promise<void> {
     }).pipe(startWith([{ addedNodes: [document.body], removedNodes: [] }]))
 
     // TODO handle subscription
-    await injectCodeIntelligence(mutations)
+    const codeHost = await determineCodeHost()
+    if (codeHost) {
+        await injectCodeIntelligenceToCodeHost(mutations, codeHost)
+    }
 }
 
 function init(): void {