+++ cmd/repo-updater/repos/migrations.go
@@ -37,7 +37,10 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 	return migrate(func(ctx context.Context, s Store) error {
 		const prefix = "migrate.repos-enabled-state-deprecation"
 
-		es, err := s.ListExternalServices(ctx, kinds...)
+		es, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
+			Kinds: kinds,
+		})
+
 		if err != nil {
 			return errors.Wrapf(err, "%s.list-external-services", prefix)
 		}
@@ -58,7 +61,10 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 			return errors.Wrapf(err, "%s.sources.list-repos", prefix)
 		}
 
-		stored, err := s.ListRepos(ctx, kinds...)
+		stored, err := s.ListRepos(ctx, StoreListReposArgs{
+			Kinds: kinds,
+		})
+
 		if err != nil {
 			return errors.Wrapf(err, "%s.store.list-repos", prefix)
 		}
@@ -200,7 +206,10 @@ func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration 
 	return migrate(func(ctx context.Context, s Store) error {
 		const prefix = "migrate.github-set-default-repository-query"
 
-		svcs, err := s.ListExternalServices(ctx, "github")
+		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
+			Kinds: []string{"github"},
+		})
+
 		if err != nil {
 			return errors.Wrapf(err, "%s.list-external-services", prefix)
 		}
@@ -252,7 +261,10 @@ func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 	return migrate(func(ctx context.Context, s Store) error {
 		const prefix = "migrate.gitlab-set-default-project-query"
 
-		svcs, err := s.ListExternalServices(ctx, "gitlab")
+		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
+			Kinds: []string{"gitlab"},
+		})
+
 		if err != nil {
 			return errors.Wrapf(err, "%s.list-external-services", prefix)
 		}
@@ -271,6 +283,55 @@ func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 			c.ProjectQuery = append(c.ProjectQuery, "?membership=true")
 
 			edited, err := jsonc.Edit(svc.Config, c.ProjectQuery, "projectQuery")
+			if err != nil {
+				return errors.Wrapf(err, "%s.edit-json", prefix)
+			}
+
+			svc.Config = edited
+			svc.UpdatedAt = now
+		}
+
+		if err = s.UpsertExternalServices(ctx, svcs...); err != nil {
+			return errors.Wrapf(err, "%s.upsert-external-services", prefix)
+		}
+
+		return nil
+	})
+}
+
+// BitbucketServerSetDefaultRepositoryQueryMigration returns a Migration that changes all
+// configurations of BitbucketServer external services to explicitly have the new
+// `repositoryQuery` setting set to a value that results in the semantically equivalent
+// behaviour of mirroring all repos accessible to the configured token.
+func BitbucketServerSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration {
+	return migrate(func(ctx context.Context, s Store) error {
+		const prefix = "migrate.bitbucketserver-set-default-repository-query"
+
+		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
+			Kinds: []string{"bitbucketserver"},
+		})
+
+		if err != nil {
+			return errors.Wrapf(err, "%s.list-external-services", prefix)
+		}
+
+		now := clock()
+		for _, svc := range svcs {
+			var c schema.BitbucketServerConnection
+			if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
+				return fmt.Errorf("%s: external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
+			}
+
+			if len(c.RepositoryQuery) != 0 {
+				continue
+			}
+
+			c.RepositoryQuery = append(c.RepositoryQuery,
+				"?visibility=private",
+				"?visibility=public",
+			)
+
+			edited, err := jsonc.Edit(svc.Config, c.RepositoryQuery, "repositoryQuery")
 			if err != nil {
 				return errors.Wrapf(err, "%s.edit-json", prefix)
 			}