+++ cmd/repo-updater/repos/syncer_test.go
@@ -3,6 +3,7 @@ package repos_test
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 	"time"
 
@@ -11,6 +12,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/cmd/repo-updater/repos"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc/bitbucketserver"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
 )
@@ -108,6 +110,23 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 		repos.Opt.RepoSources(gitlabService.URN()),
 	)
 
+	bitbucketServerService := &repos.ExternalService{
+		ID:   20,
+		Kind: "BITBUCKETSERVER",
+	}
+
+	bitbucketServerRepo := (&repos.Repo{
+		Name:     "bitbucketserver.mycorp.com/org/foo",
+		Metadata: &bitbucketserver.Repo{},
+		ExternalRepo: api.ExternalRepoSpec{
+			ID:          "23456",
+			ServiceID:   "https://bitbucketserver.mycorp.com/",
+			ServiceType: "bitbucketServer",
+		},
+	}).With(
+		repos.Opt.RepoSources(bitbucketServerService.URN()),
+	)
+
 	clock := repos.NewFakeClock(time.Now(), 0)
 
 	type testCase struct {
@@ -128,6 +147,7 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 	}{
 		{repo: githubRepo, svc: githubService},
 		{repo: gitlabRepo, svc: gitlabService},
+		{repo: bitbucketServerRepo, svc: bitbucketServerService},
 	} {
 		svcdup := tc.svc.With(repos.Opt.ExternalServiceID(tc.svc.ID + 1))
 		testCases = append(testCases,
@@ -151,6 +171,19 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 				store:   s,
 				stored: repos.Repos{tc.repo.With(func(r *repos.Repo) {
 					r.ExternalRepo.ID = ""
+				})},
+				now: clock.Now,
+				diff: repos.Diff{Modified: repos.Repos{
+					tc.repo.With(repos.Opt.RepoModifiedAt(clock.Time(1))),
+				}},
+				err: "<nil>",
+			},
+			testCase{
+				name:    "updated external_id",
+				sourcer: repos.NewFakeSourcer(nil, repos.NewFakeSource(tc.svc.Clone(), nil, tc.repo.Clone())),
+				store:   s,
+				stored: repos.Repos{tc.repo.With(func(r *repos.Repo) {
+					r.ExternalRepo.ID = "old-and-out-of-date"
 				})},
 				now: clock.Now,
 				diff: repos.Diff{Modified: repos.Repos{
@@ -256,11 +289,13 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 			},
 			func() testCase {
 				var update interface{}
-				switch tc.repo.ExternalRepo.ServiceType {
+				switch strings.ToLower(tc.repo.ExternalRepo.ServiceType) {
 				case "github":
 					update = &github.Repository{IsArchived: true}
 				case "gitlab":
 					update = &gitlab.Project{Archived: true}
+				case "bitbucketserver":
+					update = &bitbucketserver.Repo{Public: true}
 				default:
 					panic("test must be extended with new external service kind")
 				}
@@ -287,11 +322,6 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 
 	return func(t *testing.T) {
 		t.Helper()
-
-		kinds := []string{
-			"GITHUB",
-			"GITLAB",
-		}
 
 		for _, tc := range testCases {
 			tc := tc
@@ -322,7 +352,7 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 				}
 
 				syncer := repos.NewSyncer(st, tc.sourcer, nil, now)
-				diff, err := syncer.Sync(ctx, kinds...)
+				diff, err := syncer.Sync(ctx)
 
 				var want repos.Repos
 				want.Concat(diff.Added, diff.Modified, diff.Unmodified, diff.Deleted)
@@ -341,7 +371,7 @@ func testSyncerSync(s repos.Store) func(*testing.T) {
 				}
 
 				if st != nil {
-					have, _ := st.ListRepos(ctx)
+					have, _ := st.ListRepos(ctx, repos.StoreListReposArgs{})
 					for _, d := range have {
 						d.ID = 0 // Exclude auto-generated ID from comparisons
 					}