+++ shared/src/actions/ActionsNavItems.tsx
@@ -9,18 +9,34 @@ import { ActionsState } from './actions'
 import { ActionsProps } from './ActionsContainer'
 
 export interface ActionNavItemsClassProps {
+    /**
+     * CSS class name for one action item (`<button>` or `<a>`)
+     */
     actionItemClass?: string
+
+    /**
+     * Additional CSS class name when the action item is a toogle in its enabled state.
+     */
     actionItemPressedClass?: string
+
+    actionItemIconClass?: string
+
+    /**
+     * CSS class name for each `<li>` element wrapping the action item.
+     */
     listItemClass?: string
 }
 
-interface Props extends ActionsProps, ActionNavItemsClassProps, TelemetryProps {
+export interface ActionsNavItemsProps extends ActionsProps, ActionNavItemsClassProps, TelemetryProps {
     /**
      * If true, it renders a `<ul className="nav">...</ul>` around the items. If there are no items, it renders `null`.
      *
      * If falsey (the default behavior), it emits a fragment of just the `<li>`s.
      */
     wrapInList?: boolean
+    /**
+     * Only applied if `wrapInList` is `true`
+     */
 
     listClass?: string
 }
@@ -29,7 +45,7 @@ interface Props extends ActionsProps, ActionNavItemsClassProps, TelemetryProps {
  * Renders the actions as a fragment of <li class="nav-item"> elements, for use in a Bootstrap <ul
  * class="nav"> or <ul class="navbar-nav">.
  */
-export class ActionsNavItems extends React.PureComponent<Props, ActionsState> {
+export class ActionsNavItems extends React.PureComponent<ActionsNavItemsProps, ActionsState> {
     public state: ActionsState = {}
 
     private scopeChanges = new Subject<ContributionScope | undefined>()
@@ -60,21 +76,23 @@ export class ActionsNavItems extends React.PureComponent<Props, ActionsState> {
         }
 
         const actionItems = getContributedActionItems(this.state.contributions, this.props.menu).map((item, i) => (
-            <li key={i} className={this.props.listItemClass || 'nav-item'}>
-                <ActionItem
-                    key={i}
-                    {...item}
-                    {...this.props}
-                    variant="actionItem"
-                    className={this.props.actionItemClass}
-                    pressedClassName={this.props.actionItemPressedClass}
-                />
-            </li>
+            <React.Fragment key={item.action.id}>
+                {' '}
+                <li className={this.props.listItemClass}>
+                    <ActionItem
+                        key={item.action.id}
+                        {...item}
+                        {...this.props}
+                        variant="actionItem"
+                        iconClassName={this.props.actionItemIconClass}
+                        className={this.props.actionItemClass}
+                        pressedClassName={this.props.actionItemPressedClass}
+                    />
+                </li>
+            </React.Fragment>
         ))
         if (this.props.wrapInList) {
-            return actionItems.length > 0 ? (
-                <ul className={`nav ${this.props.listClass || ''}`}>{actionItems}</ul>
-            ) : null
+            return actionItems.length > 0 ? <ul className={this.props.listClass}>{actionItems}</ul> : null
         }
         return <>{actionItems}</>
     }